fix(Scrolling Chat Container): move layout to flex based
currently, the height of the #scrollingChat selector is calc'd for 100% height against a fixed 4rem
for the top margin offset. this causes issues if content besides what is accounted for in the 4rem
is added. The more flexible approach is to convert .chat-column to a flex container, and allow
#scrollingChat to flex: 1, which will have it take up all available space that isn't used by the
other elements.